Teach magit how to PR and fork
Add a new popup menu to magit, under the 'h' key. This popup provides
access to `hub`, the Github CLI.

https://github.com/github/hub

The Github popup supports creating pull requests (`p`) and
forks (`f`). After creating a pull request, the URL for the request will
be stored in the kill ring for easy access.

`hub` does not recognize the `--literal-pathspec` flag that magit passes
to git, so the following PR is also needed (otherwise `hub` falls back
to invoking `git`, and `git` doesn't know about `pull-request` or `fork`):

https://github.com/github/hub/pull/942