setup: fix PATH to use DST instead of SRC
The PATH was incorrectly referencing $SRC/.local/bin, which caused
lua-shimlink and other scripts to not be found during bootstrap.
Since the dotfiles are checked out to $DST ($HOME) by the _git()
function, the PATH should reference $DST/.local/bin where the
lua-shimlink symlink and other scripts will be located.