home: fix manifest generation with proper find -exec
The find -exec command was using "{}" as a literal string in sh -c,
causing stat to fail and filenames to be misinterpreted as shell
commands. This created spurious directories with names like
"config.draw.predicate" and "opts.directory".

Fixed by passing the filename as a positional argument ($1) to the
shell command instead.